refactor(playerbot): Remove ~18,700 lines of dead/duplicate code
Phase 1-2 of enterprise code audit cleanup:

Deleted backup files (~15,000 lines):
- 11 .bak/.orig files removed (untracked)

Deleted duplicate systems (~2,734 lines):
- Game/FlightMasterManager.* (duplicate of Interaction/)
- Game/NPCInteractionManager.* (duplicate of Interaction/Core/)

Deleted orphaned stubs (~84 lines):
- Core/Managers/BotPriorityManager.cpp (real: Session/)
- Core/Managers/BotWorldSessionMgr.cpp (real: Session/)
- Group/LFGBotManager.cpp (real: LFG/)
- Group/LFGBotSelector.cpp (real: LFG/)
- Group/LFGGroupCoordinator.cpp (real: LFG/)
- Advanced/InstanceCoordination.cpp (real: Dungeon/)

Deleted legacy/deprecated files (~860 lines):
- AI/ClassAI/ClassAI_Legacy.h
- AI/ClassAI/ClassAI_Refactored.cpp
- AI/ClassAI/CooldownManager.h (redirect file)
- AI/PositionManager.h (orphaned header)

Updated include paths:
- CooldownManager.h -> Common/CooldownManager.h
- Game/FlightMasterManager.h -> Interaction/FlightMasterManager.h
- Removed duplicate includes in HavocDemonHunter.h, DisciplinePriest.h

Updated CMakeLists.txt to remove deleted file references.
