--- src/modules/Playerbot/AI/ClassAI/CooldownManager.h
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2025 TrinityCore <https://www.trinitycore.org/>
- *
- * This file has been replaced by the unified cooldown system.
- * Include the new location instead.
- */
-
-#pragma once
-
-// Redirect to unified cooldown management system
-#include "Common/CooldownManager.h"
