+++ src/modules/Playerbot/AI/ClassAI/Rogues/RogueAI.cpp
@@ -27,7 +27,7 @@
 #include "../Combat/TargetSelector.h"
 #include "../Combat/PositionManager.h"
 #include "../Combat/InterruptManager.h"
-#include "../CooldownManager.h"
+#include "../Common/CooldownManager.h"
 #include "CellImpl.h"
 #include "GridNotifiersImpl.h"
 #include "Log.h"
