+++ src/modules/Playerbot/AI/ClassAI/Rogues/RogueAI.h
@@ -18,7 +18,7 @@
 #include "../Combat/TargetSelector.h"
 #include "../Combat/PositionManager.h"
 #include "../Combat/InterruptManager.h"
-#include "../CooldownManager.h"
+#include "../Common/CooldownManager.h"
 #include <memory>
 #include <atomic>
 #include <chrono>
