--- src/modules/Playerbot/Core/Managers/BotWorldSessionMgr.cpp
@@ -1,14 +1,0 @@
-#include "BotWorldSessionMgr.h"
-
-namespace Playerbot
-{
-
-BotWorldSessionMgr::BotWorldSessionMgr(Player* bot) : _bot(bot)
-{
-    if (!_bot)
-        TC_LOG_ERROR("playerbot", "BotWorldSessionMgr: null bot!");
-}
-
-BotWorldSessionMgr::~BotWorldSessionMgr() {}
-
-} // namespace Playerbot
