--- src/modules/Playerbot/Game/FlightMasterManager.h
@@ -1,348 +1,0 @@
-/*
- * Copyright (C) 2024+ TrinityCore <http://www.trinitycore.org/>
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the
- * Free Software Foundation; either version 2 of the License, or (at your
- * option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-#ifndef _PLAYERBOT_FLIGHT_MASTER_MANAGER_H
-#define _PLAYERBOT_FLIGHT_MASTER_MANAGER_H
-
-#include "Common.h"
-#include "ObjectGuid.h"
-#include "Position.h"
-#include <optional>
-#include <vector>
-
-class Player;
-class Creature;
-
-namespace Playerbot
-{
-    /**
-     * @enum FlightResult
-     * @brief Result codes for flight master operations
-     */
-    enum class FlightResult
-    {
-        SUCCESS = 0,                    ///< Flight activated successfully
-        FLIGHT_MASTER_NOT_FOUND,        ///< Flight master NPC not found
-        NOT_A_FLIGHT_MASTER,            ///< NPC is not a flight master
-        OUT_OF_RANGE,                   ///< Too far from flight master
-        NODE_UNKNOWN,                   ///< Destination taxi node not discovered
-        PATH_NOT_FOUND,                 ///< No taxi path to destination
-        INSUFFICIENT_GOLD,              ///< Cannot afford flight cost
-        ALREADY_IN_FLIGHT,              ///< Player already flying
-        PLAYER_INVALID,                 ///< Player is null or invalid
-        DESTINATION_INVALID,            ///< Invalid destination coordinates
-        SAME_LOCATION                   ///< Already at destination
-    };
-
-    /**
-     * @enum FlightPathStrategy
-     * @brief Strategy for selecting flight path when multiple routes exist
-     */
-    enum class FlightPathStrategy
-    {
-        SHORTEST_DISTANCE,              ///< Select path with shortest total distance
-        FEWEST_STOPS,                   ///< Select path with fewest intermediate stops
-        LOWEST_COST,                    ///< Select cheapest flight path
-        FASTEST_TIME                    ///< Select path with shortest flight time
-    };
-
-    /**
-     * @struct FlightPathInfo
-     * @brief Information about a taxi flight path
-     */
-    struct FlightPathInfo
-    {
-        uint32 pathId;                  ///< Taxi path ID
-        uint32 sourceNode;              ///< Source taxi node ID
-        uint32 destinationNode;         ///< Destination taxi node ID
-        ::std::vector<uint32> nodes;      ///< All nodes in path
-        float totalDistance;            ///< Total flight distance (yards)
-        uint32 flightTime;              ///< Estimated flight time (seconds)
-        uint32 goldCost;                ///< Flight cost in copper
-        uint32 stopCount;               ///< Number of intermediate stops
-
-        FlightPathInfo()
-            : pathId(0), sourceNode(0), destinationNode(0),
-              totalDistance(0.0f), flightTime(0), goldCost(0), stopCount(0) {}
-    };
-
-    /**
-     * @struct FlightMasterLocation
-     * @brief Information about a flight master NPC
-     */
-    struct FlightMasterLocation
-    {
-        ObjectGuid guid;                ///< Flight master NPC GUID
-        uint32 entry;                   ///< Creature entry ID
-        Position position;              ///< Location coordinates
-        uint32 taxiNode;                ///< Associated taxi node ID
-        float distanceFromPlayer;       ///< Distance from player (yards)
-        ::std::string name;               ///< Flight master name
-
-        FlightMasterLocation()
-            : entry(0), taxiNode(0), distanceFromPlayer(0.0f) {}
-    };
-
-    /**
-     * @class FlightMasterManager
-     * @brief High-performance flight master system for bot travel
-     *
-     * Purpose:
-     * - Automate taxi flight usage for bot long-distance travel
-     * - Find nearest flight masters
-     * - Select optimal flight paths based on strategy
-     * - Integrate with TrinityCore's taxi system (Player::ActivateTaxiPathTo)
-     *
-     * Features:
-     * - Automatic flight master detection (UNIT_NPC_FLAG_FLIGHTMASTER)
-     * - Multi-path route finding (supports multiple hops)
-     * - Strategy-based path selection (distance, cost, time, stops)
-     * - Taxi node discovery validation
-     * - Gold cost calculation with reputation discounts
-     *
-     * Performance Targets:
-     * - Flight master search: < 1ms (map creature iteration)
-     * - Path calculation: < 2ms (graph traversal)
-     * - Flight activation: < 1ms (TrinityCore API call)
-     *
-     * Quality Standards:
-     * - NO shortcuts - Full Player::ActivateTaxiPathTo integration
-     * - Complete error handling and validation
-     * - Production-ready code (no TODOs)
-     *
-     * @code
-     * // Example usage:
-     * FlightMasterManager mgr;
-     *
-     * // Find nearest flight master
-     * auto flightMaster = mgr.FindNearestFlightMaster(bot);
-     * if (flightMaster)
-     * {
-     *     TC_LOG_DEBUG("playerbot", "Nearest flight master: {} at {:.1f} yards",
-     *                  flightMaster->name, flightMaster->distanceFromPlayer);
-     * }
-     *
-     * // Fly to destination
-     * Position destination(1000.0f, 2000.0f, 50.0f);
-     * FlightResult result = mgr.FlyToPosition(bot, destination, FlightPathStrategy::SHORTEST_DISTANCE);
-     * @endcode
-     */
-    class FlightMasterManager
-    {
-    public:
-        FlightMasterManager() = default;
-        ~FlightMasterManager() = default;
-
-        /// Delete copy/move (stateless utility class)
-        FlightMasterManager(FlightMasterManager const&) = delete;
-        FlightMasterManager(FlightMasterManager&&) = delete;
-        FlightMasterManager& operator=(FlightMasterManager const&) = delete;
-        FlightMasterManager& operator=(FlightMasterManager&&) = delete;
-
-        /**
-         * @brief Activates taxi flight to destination position
-         *
-         * Workflow:
-         * 1. Find nearest flight master to player
-         * 2. Find nearest taxi node to destination
-         * 3. Calculate flight path using TaxiPathGraph
-         * 4. Validate player knows all nodes in path
-         * 5. Activate flight via Player::ActivateTaxiPathTo()
-         *
-         * @param player Bot taking flight
-         * @param destination Target position
-         * @param strategy Path selection strategy
-         * @return Result code indicating success or failure reason
-         *
-         * Performance: < 5ms total (1ms search + 2ms path + 1ms activation)
-         * Thread-safety: Main thread only (modifies player state)
-         */
-        [[nodiscard]] FlightResult FlyToPosition(
-            Player* player,
-            Position const& destination,
-            FlightPathStrategy strategy = FlightPathStrategy::SHORTEST_DISTANCE);
-
-        /**
-         * @brief Activates taxi flight to specific taxi node
-         *
-         * @param player Bot taking flight
-         * @param destinationNodeId Target taxi node ID
-         * @param strategy Path selection strategy
-         * @return Result code indicating success or failure reason
-         *
-         * Performance: < 4ms (1ms search + 2ms path + 1ms activation)
-         * Thread-safety: Main thread only
-         */
-        [[nodiscard]] FlightResult FlyToTaxiNode(
-            Player* player,
-            uint32 destinationNodeId,
-            FlightPathStrategy strategy = FlightPathStrategy::SHORTEST_DISTANCE);
-
-        /**
-         * @brief Finds nearest flight master to player
-         *
-         * @param player Player reference point
-         * @param maxDistance Maximum search distance (0 = unlimited)
-         * @return Flight master location, or std::nullopt if none found
-         *
-         * Performance: O(n) where n = creatures on map, ~1ms typical
-         * Thread-safety: Thread-safe (read-only map access)
-         */
-        [[nodiscard]] static ::std::optional<FlightMasterLocation> FindNearestFlightMaster(
-            Player const* player,
-            float maxDistance = 0.0f);
-
-        /**
-         * @brief Finds nearest taxi node to position (faction-aware)
-         *
-         * @param position Target position
-         * @param mapId Map to search on
-         * @param player Player for faction filtering (optional, nullptr = no faction check)
-         * @return Nearest taxi node ID for player's faction, or 0 if none found
-         *
-         * Note: When player is provided, only returns taxi nodes that match
-         * the player's faction (using TaxiNodeFlags::ShowOnHordeMap/ShowOnAllianceMap)
-         *
-         * Performance: O(n) where n = taxi nodes, ~0.5ms typical
-         * Thread-safety: Thread-safe (read-only DBC access)
-         */
-        [[nodiscard]] static uint32 FindNearestTaxiNode(
-            Position const& position,
-            uint32 mapId,
-            Player const* player = nullptr);
-
-        /**
-         * @brief Calculates taxi flight path between two nodes
-         *
-         * Uses TaxiPathGraph to find shortest path through taxi network.
-         * Validates player has discovered all nodes in path.
-         *
-         * @param player Player (for node discovery check)
-         * @param sourceNode Source taxi node ID
-         * @param destinationNode Destination taxi node ID
-         * @param strategy Path selection strategy
-         * @return Flight path info, or std::nullopt if no path found
-         *
-         * Performance: O(V + E) graph traversal, ~2ms typical
-         * Thread-safety: Thread-safe (read-only graph access)
-         */
-        [[nodiscard]] static ::std::optional<FlightPathInfo> CalculateFlightPath(
-            Player const* player,
-            uint32 sourceNode,
-            uint32 destinationNode,
-            FlightPathStrategy strategy);
-
-        /**
-         * @brief Checks if player has discovered taxi node
-         *
-         * @param player Player to check
-         * @param nodeId Taxi node ID
-         * @return true if player knows this taxi node
-         *
-         * Performance: < 0.01ms (bitmask check)
-         * Thread-safety: Thread-safe (read-only)
-         */
-        [[nodiscard]] static bool HasTaxiNode(
-            Player const* player,
-            uint32 nodeId);
-
-        /**
-         * @brief Gets human-readable error message for flight result
-         *
-         * @param result Flight result code
-         * @return String description
-         */
-        [[nodiscard]] static char const* GetResultString(FlightResult result);
-
-        /**
-         * @brief Checks if a valid taxi path exists between two nodes
-         *
-         * Uses TaxiPathGraph to verify a route exists. This should be called
-         * BEFORE creating taxi legs to prevent invalid cross-continent routes.
-         *
-         * @param startNode Source taxi node ID
-         * @param endNode Destination taxi node ID
-         * @param player Player for faction/condition checks
-         * @return true if a valid path exists, false otherwise
-         *
-         * Performance: O(V + E) graph traversal, ~2ms typical
-         * Thread-safety: Thread-safe (read-only graph access)
-         */
-        [[nodiscard]] static bool HasValidFlightPath(
-            uint32 startNode,
-            uint32 endNode,
-            Player const* player);
-
-    private:
-        /**
-         * @brief Validates flight activation preconditions
-         *
-         * Checks:
-         * - Player is valid and not already flying
-         * - Source and destination nodes are valid
-         * - Player has discovered destination node
-         * - Player has sufficient gold
-         *
-         * @param player Player attempting flight
-         * @param sourceNode Source taxi node
-         * @param destinationNode Destination taxi node
-         * @return Result code (SUCCESS if valid, error otherwise)
-         */
-        [[nodiscard]] static FlightResult ValidateFlight(
-            Player const* player,
-            uint32 sourceNode,
-            uint32 destinationNode);
-
-        /**
-         * @brief Calculates total flight cost for path
-         *
-         * @param player Player taking flight
-         * @param nodes Taxi node path
-         * @return Total gold cost in copper
-         *
-         * Performance: < 0.1ms
-         */
-        [[nodiscard]] static uint32 CalculateFlightCost(
-            Player const* player,
-            ::std::vector<uint32> const& nodes);
-
-        /**
-         * @brief Estimates flight time for path
-         *
-         * @param nodes Taxi node path
-         * @return Estimated flight time in seconds
-         *
-         * Performance: < 0.1ms
-         */
-        [[nodiscard]] static uint32 EstimateFlightTime(
-            ::std::vector<uint32> const& nodes);
-
-        /**
-         * @brief Calculates total flight distance
-         *
-         * @param nodes Taxi node path
-         * @return Total distance in yards
-         *
-         * Performance: < 0.1ms
-         */
-        [[nodiscard]] static float CalculateFlightDistance(
-            ::std::vector<uint32> const& nodes);
-    };
-
-} // namespace Playerbot
-
-#endif // _PLAYERBOT_FLIGHT_MASTER_MANAGER_H
